Added windows forms controls and created bindings to the form code
For right now, any information you ask the player to enter, just save it as a variable. Don't worry about the database part yet. On the third page, if you'd like, you could actually create labels or some other container that displays those variables in real time. But that is definitely not going to be what we end up going with.

When starting on any of this, look up private accessors first. Don't just make this stuff public strings and whatever. Make them private and "get; set;" them through properties. If you don't know what that means, read up on those private accessors in C# and setting properties of a private variable.